% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/carto_hcl.R
\name{carto_hcl}
\alias{carto_hcl}
\title{CARTO Sequential and Diverging Palettes Emulated by HCL Palettes}
\usage{
carto_hcl(n, palette = "Geyser", ..., rev = FALSE, h1, h2, h3, c1, c2, c3,
  l1, l2, l3, p1, p2, p3, p4)
}
\arguments{
\item{n}{the number of colors (\eqn{\ge 1}{>= 1}) to be in the palette.}

\item{palette}{character with the name (see details).}

\item{\dots}{arguments passed to \code{\link{sequential_hcl}}.}

\item{rev}{logical. Should the palette be reversed?}

\item{h1}{numeric. Starting hue coordinate.}

\item{h2}{numeric. Either ending hue coordinate (for sequential palettes)
or center hue coordinate (for diverging palettes).}

\item{h3}{numeric. Either \code{NA} (for sequential palettes) or ending hue
coordinate (for diverging palettes).}

\item{c1}{numeric. Chroma coordinate corresponding to \code{h1}.}

\item{c2}{numeric. Chroma coordinate corresponding to \code{h2}.}

\item{c3}{numeric. Chroma coordinate corresponding to \code{h3} (if any).}

\item{l1}{numeric. Luminance coordinate corresponding to \code{h1}.}

\item{l2}{numeric. Luminance coordinate corresponding to \code{h2}.}

\item{l3}{numeric. Luminance coordinate corresponding to \code{h3} (if any).}

\item{p1}{numeric. Power parameter for chroma coordinates in (first) sequential palette.}

\item{p2}{numeric. Power parameter for luminance coordinates in (first) sequential palette.}

\item{p3}{numeric. Power parameter for chroma coordinates in second sequential palette for diverging palettes.}

\item{p4}{numeric. Power parameter for luminance coordinates in second sequential palette for diverging palettes.}
}
\value{
A character vector with (s)RGB codings of the colors in the palette.
}
\description{
Sequential and diverging CARTO color palettes, emulated by sequential
HCL palettes (combining two sequential palettes for diverging palettes).
}
\details{
The \code{carto_hcl} function simply calls \code{\link{sequential_hcl}}
with a prespecified set of hue/chorma/luminance parameters that have
been chosen such that the CARTO color palettes (\url{https://carto.com/carto-colors/})
are approximately emulated. In case of diverging CARTO palettes two
sequential HCL palettes are combined, possibly with rather unbalanced
chroma/luminance specifications.

Available sequential palettes: RedOr, OrYel, Peach, PinkYl, Mint, BluGrn,
DarkMint, Emrld, ag_GrnYl, BluYl, Teal, TealGrn, PurpOr, Sunset, Magenta.

Available diverging palettes: ArmyRose, Fall, Geyser, Temps, Tropic, Earth.

Furthermore some CARTO palettes cannot be emulated well with sequential
HCL palettes, typically due to requiring non-monotonic chroma sequences
or sometimes nonlinear hue sequences. The CARTO palettes not available
in \code{carto_hcl} are: Burg, BurgYl, Purp, SunsetDark, ag_Sunset, BrwnYl, TealRose.
}
\examples{
## show emulated CARTO palettes
carto_pal <- function(palette, border = "light gray")
{
  n <- 7
  col <- carto_hcl(n, palette)
  plot(0, 0, type="n", xlim = c(0, 1), ylim = c(0, 1), xaxs = "i", axes = FALSE, xlab = "", ylab = "")
  rect(0:(n-1)/n, 0, 1:n/n, 1, col = col, border = border)
  axis(2, at = 0.5, tick = FALSE, labels = palette)
}
carto_seq <- c("RedOr", "OrYel", "Peach", "PinkYl", "Mint", "BluGrn", "DarkMint",
  "Emrld", "ag_GrnYl", "BluYl", "Teal", "TealGrn", "PurpOr", "Sunset", "Magenta")
carto_div <- c("ArmyRose", "Earth", "Fall", "Geyser", "Temps", "Tropic")

## sequential palettes
par(mar = c(0, 5.5, 0, 0.5), mfrow = c(8, 2), las = 1)
for(n in carto_seq) carto_pal(n)

## diverging palettes
par(mar = c(0, 5.5, 0, 0.5), mfrow = c(6, 1), las = 1)
for(n in carto_div) carto_pal(n)

## compared to diverge_hcl() the diverging CARTO palettes are typically warmer
## but also less balanced with respect to chroma/luminance, see e.g.,
specplot(carto_hcl(7, "Fall"), rgb = FALSE)
}
\seealso{
\code{\link[colorspace]{sequential_hcl}}
}
\keyword{color}
