% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hcl_converter.R
\name{check_image_type}
\alias{check_image_type}
\title{Check Image Type based on Postfix}
\usage{
check_image_type(x)
}
\arguments{
\item{x, }{\code{string} containing (full) path to image.}
}
\value{
Returns a \code{list} with two elements. Each can take
   \code{TRUE} or \code{FALSE}. If image is of type \code{png}
   \code{png=TRUE}, if image is of type \code{jpg/jpeg} \code{jpg=TRUE}.
   If non of both, both will be \code{FALSE}. Method is not case sensitive.
}
\description{
Checking image file type based on image file name! Used to decide
which package has to be used to read an image from disc (\code{png}/\code{jpeg}).
}
