% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hcl_converter.R
\name{convert_image}
\alias{convert_image}
\title{Convert Colors of an Image}
\usage{
convert_image(img, type, target)
}
\arguments{
\item{img}{\code{array} as returned by \code{readPNG} and \code{readJPEG}
of size \code{height x width x depth}. The depth coordinate contains
R/G/B and alpha if given (png).}

\item{type}{\code{string} name of the function which will be used to
convert the colors (\code{deutan}, \code{protan}, \code{tritan}, \code{desaturate}).
If set to \code{original} the image will be written as is.}

\item{target}{\code{string} with (full) path to resulting image. Has to
be a png image name!}
}
\description{
Used in \code{hcl_converter}. Takes an image object and converts
the colors using deutan/protan/tritan/desaturate functions. Image
will be written to disc as a png image.
}
