% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/colorspace.R
\name{coords}
\alias{coords}
\title{Extract the Numerical Coordinates of a Color}
\usage{
coords(color)
}
\arguments{
\item{color}{A color.}
}
\value{
A numeric matrix giving the coordinates of the color.
}
\description{
This function returns a matrix with three columns which give the coordinates
of a color in its natural color space.
}
\examples{
x <- RGB(1, 0, 0)
coords(as(x, "HSV"))
}
\seealso{
\code{\link{RGB}}, \code{\link{XYZ}}, \code{\link{LAB}},
\code{\link{polarLAB}}, \code{\link{LUV}}, \code{\link{polarLUV}},
\code{\link{mixcolor}}.
}
\author{
Ross Ihaka
}
\keyword{color}
