% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cvd_image.R
\name{cvd_emulator}
\alias{cvd_emulator}
\title{Graphical User Interface to Check Images for Color Constraints}
\usage{
cvd_emulator(file, overwrite = FALSE, shiny.trace = FALSE)
}
\arguments{
\item{file}{If not set, an interactive GUI will be started. If \code{x} is of type
\code{character} it has to be the full path to an image of type png or jpg/jpeg.
The image will be converted and stored on disc, no GUI.}

\item{overwrite}{\code{logical}. Only used if \code{file} is provided. Allow the
function to overwrite files on disc if they exist.}

\item{shiny.trace}{\code{logical}. Can be set to \code{TRUE} for more verbose
output when the GUI is started (development flag).}
}
\description{
A graphical user interface (GUI) to check an existing jpg/png image for (possible)
color constraints. The image will be converted to protanope vision, deuteranope vision,
and a desaturated version (monochromatic vision).
Allows a rapid check whether the colors used in the image show some
constraints with respect to color deficiency or color blindness.
}
\author{
Reto Stauffer, Claus O. Wilke, Achim Zeileis
}
