% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/colorspace.R
\name{getWhite}
\alias{getWhite}
\alias{setWhite}
\alias{whitepoint}
\alias{setWhite}
\title{Access or Modify the Whitepoint}
\usage{
getWhite()

setWhite(white)
}
\arguments{
\item{white}{A color object as generated by \code{XYZ} or \code{NULL},
which means use CIE D65.}
}
\value{
\code{getWhite} returns the current whitepoint.
}
\description{
This function can be used to control the single global whitepoint
that affects all colour conversions.
}
\examples{
getWhite()
# Illuminant E 
setWhite(XYZ(100, 100, 100))
}
\seealso{
\code{\link{XYZ}} and \link{color-class}.
}
\author{
Paul Murrell
}
\keyword{color}
