% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hcl_converter.R
\name{hcl_converter}
\alias{hcl_converter}
\title{Graphical User Interface to Check Images for Color Constriants}
\usage{
hcl_converter(x, overwrite = FALSE, shiny.trace = FALSE)
}
\arguments{
\item{shiny.trace, }{\code{logical}. Can be set to \code{TRUE} for more verbose
output (development).}
}
\description{
A graphical user interface (GUI) to check an existing jpg/png image for (possible)
color constraints. The image will be converted to protanope vision, deutanope vision,
and a desaturated version (monochromatic vision).
Allows a rapid check whether the colors used in the image show some
constraints with respect to color deficiency or color blindness.
}
