% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hcl_palettes.R
\name{hcl_palettes}
\alias{hcl_palettes}
\alias{print.hcl_palettes}
\alias{summary.hcl_palettes}
\alias{plot.hcl_palettes}
\alias{qualitative_hcl}
\title{HCL Color Palettes}
\usage{
hcl_palettes(type = NULL, name = NULL, plot = FALSE, n = 5L, ...)

\method{print}{hcl_palettes}(x, ...)

\method{summary}{hcl_palettes}(object, ...)

\method{plot}{hcl_palettes}(x, n = 5L, fixup = TRUE, ...)

qualitative_hcl(n, h = c(0, 360 * (n - 1)/n), c. = 50, l = 70,
  fixup = TRUE, alpha = 1, palette = NULL, rev = FALSE, ..., h1, h2, c1,
  l1)
}
\description{
Qualitative, sequential (single-hue and multi-hue), and diverged
color palettes based on the HCL (hue-chroma-luminance) color model.
}
\details{
Details are still tbd...
}
\examples{
## overview of all _named_ HCL palettes
hcl_palettes()

## visualize
hcl_palettes("qualitative", plot = TRUE)
hcl_palettes("sequential (single-hue)", plot = TRUE)
hcl_palettes("sequential (multi-hue)", plot = TRUE)
hcl_palettes("diverging", plot = TRUE)

## inspect a specific palette
## (upper-case, spaces, etc. are ignored for matching)
hcl_palettes(name = "Dark 2")
hcl_palettes(name = "dark2")

## set up actual colors
qualitative_hcl(4, h = c(0, 288), c = 50, l = 60) ## by hand
qualitative_hcl(4, palette = "dark2")             ## by name
qualitative_hcl(4, palette = "dark2", c = 80)     ## by name plus modification
}
\keyword{color}
