% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulate_cvd.R
\name{protan_transform}
\alias{protan_transform}
\title{Generate color transformation matrices using the protan model of colorblindness}
\usage{
protan_transform(sev = 1)
}
\arguments{
\item{sev}{Severity of the color vision defect, a number between 0 and 1}
}
\description{
Generate color transformation matrices using the protan model of colorblindness
}
\author{
Claire D. McWhite claire.d.mcwhite@gmail.com

Claus O. Wilke wilke@austin.utexas.edu
}
