% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scale_sequential.R
\name{scale_colour_continuous_sequential}
\alias{scale_colour_continuous_sequential}
\alias{scale_color_continuous_sequential}
\alias{scale_fill_continuous_sequential}
\title{HCL-based continuous sequential color scales for ggplot2}
\usage{
scale_colour_continuous_sequential(palette = NULL, c1 = NULL, c2 = NULL,
  l1 = NULL, l2 = NULL, h1 = NULL, h2 = NULL, p1 = NULL, p2 = NULL,
  rev = FALSE, begin = 0, end = 1, na.value = "grey50",
  guide = "colourbar", aesthetics = "colour", n_interp = 11, ...)

scale_color_continuous_sequential(...)

scale_fill_continuous_sequential(...)
}
\arguments{
\item{palette}{The name of the palette to be used. Run \code{hcl_palettes(type = "seq")} for available options.}

\item{c1}{Beginning chroma value.}

\item{c2}{Ending chroma value.}

\item{l1}{Beginning luminance value.}

\item{l2}{Ending luminance value.}

\item{h1}{Beginning hue value.}

\item{h2}{Ending hue value. If set to \code{NA}, generates a single-hue scale.}

\item{p1}{Control parameter determining how chroma should vary (1 = linear, 2 = quadratic, etc.).}

\item{p2}{Control parameter determining how luminance should vary (1 = linear, 2 = quadratic, etc.).}

\item{rev}{If \code{TRUE}, reverses the order of the colors in the color scale.}

\item{begin}{Number in the range of \code{[0, 1]} indicating to which point in the color scale the smallest data value should be mapped.}

\item{end}{Number in the range of \code{[0, 1]} indicating to which point in the color scale the largest data value should be mapped.}

\item{na.value}{Color to be used for missing data points.}

\item{guide}{Type of legend. Use \code{"colourbar"} for continuous color bar.}

\item{aesthetics}{The ggplot2 aesthetics to which this scale should be applied.}

\item{n_interp}{Number of discrete colors that should be used to interpolate the continuous color scale. 11 will work fine in most cases.}

\item{...}{common continuous scale parameters: `name`, `breaks`, `labels`, and `limits`. See
\code{\link[ggplot2]{continuous_scale}} for more details.}
}
\description{
Continuous ggplot2 color scales using the color palettes generated by \code{\link{sequential_hcl}}.
}
\details{
If both a valid palette name and palette parameters are provided then the provided palette parameters overwrite the parameters in the
named palette. This enables easy customization of named palettes.
}
\examples{
library(ggplot2)

# base plot
gg <- ggplot(iris, aes(x = Species, y = Sepal.Width, color = Sepal.Length)) + 
  geom_jitter(width = 0.3) + theme_minimal()

# default settings
gg + scale_color_continuous_sequential()

# switch palette and overwrite some default values
gg + scale_color_continuous_sequential(palette = "Reds", l1 = 20, c2 = 70, p1 = 1)

# select a range out of the entire palette
gg + scale_color_continuous_sequential(palette = "Heat", begin = .2, end = .8)

# volcano plot
nx = 87
ny = 61
df <- data.frame(height = c(volcano), x = rep(1:nx, ny), y = rep(1:ny, each = nx))
ggplot(df, aes(x, y, fill=height)) + 
  geom_raster() + scale_fill_continuous_sequential(palette = "Terrain") +
  coord_fixed(expand = FALSE)
}
