% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scale_sequential.R
\name{scale_colour_discrete_sequential}
\alias{scale_colour_discrete_sequential}
\alias{scale_color_discrete_sequential}
\alias{scale_fill_discrete_sequential}
\title{HCL-based discrete sequential color scales for ggplot2}
\usage{
scale_colour_discrete_sequential(palette = NULL, c1 = NULL, c2 = NULL,
  l1 = NULL, l2 = NULL, h1 = NULL, h2 = NULL, p1 = NULL, p2 = NULL,
  alpha = 1, rev = FALSE, nmax = NULL, order = NULL,
  aesthetics = "colour", ...)

scale_color_discrete_sequential(...)

scale_fill_discrete_sequential(...)
}
\arguments{
\item{palette}{The name of the palette to be used. Run \code{hcl_palettes(type = "seq")} for available options.}

\item{c1}{Beginning chroma value.}

\item{c2}{Ending chroma value.}

\item{l1}{Beginning luminance value.}

\item{l2}{Ending luminance value.}

\item{h1}{Beginning hue value.}

\item{h2}{Ending hue value. If set to \code{NA}, generates a single-hue scale.}

\item{p1}{Control parameter determining how chroma should vary (1 = linear, 2 = quadratic, etc.).}

\item{p2}{Control parameter determining how luminance should vary (1 = linear, 2 = quadratic, etc.).}

\item{alpha}{Numeric vector of values in the range \code{[0, 1]} for alpha transparency channel (0 means transparent and 1 means opaque).}

\item{rev}{If \code{TRUE}, reverses the order of the colors in the color scale.}

\item{nmax}{Maximum number of different colors the palette should contain. If not provided, is calculated automatically
from the data.}

\item{order}{Numeric vector listing the order in which the colors should be used. Default is \code{1:nmax}.}

\item{aesthetics}{The ggplot2 aesthetics to which this scale should be applied.}

\item{...}{common discrete scale parameters: \code{name}, \code{breaks}, \code{labels}, \code{na.value}, \code{limits} and \code{guide}. See
\code{\link[ggplot2]{discrete_scale}} for more details.}
}
\description{
Discrete ggplot2 color scales using the color palettes generated by \code{\link{sequential_hcl}}.
}
\details{
If both a valid palette name and palette parameters are provided then the provided palette parameters overwrite the parameters in the
named palette. This enables easy customization of named palettes.
}
\examples{
library(ggplot2)

# default colors
ggplot(iris, aes(Sepal.Length, Sepal.Width, color = Species)) +
  geom_point() + scale_color_discrete_sequential() + theme_classic()

# customization of named palette  
ggplot(iris, aes(Sepal.Length, Sepal.Width, color = Species)) +
  geom_point() + scale_colour_discrete_sequential(palette = "Reds", nmax = 4, p2 = 1.5) +
  theme_classic()
 
# color scale "Terrain"
ggplot(iris, aes(Sepal.Length, fill = Species)) +
  geom_density(alpha = 0.7) + scale_fill_discrete_sequential(palette = "Terrain") + theme_minimal()
}
