% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hcl_converter.R
\name{hcl_converter}
\alias{hcl_converter}
\title{Graphical User Interface to Check Images for Color Constriants}
\usage{
hcl_converter(x, overwrite = FALSE, shiny.trace = FALSE)
}
\arguments{
\item{x}{If not set an interactive GUI will be started. If \code{x} is of type
\code{character} it has to be the full path to an image of type png or jpg/jpeg.
The image will be converted and stored on disc, no GUI.}

\item{overwrite}{\code{logical}. Only used if \code{x} is provided. Allow the
function to overwrite files on disc if they exist.}

\item{shiny.trace}{\code{logical}. Can be set to \code{TRUE} for more verbose
output when the GUI is started (development flag).}
}
\description{
A graphical user interface (GUI) to check an existing jpg/png image for (possible)
color constraints. The image will be converted to protanope vision, deutanope vision,
and a desaturated version (monochromatic vision).
Allows a rapid check whether the colors used in the image show some
constraints with respect to color deficiency or color blindness.
}
