% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hcl_palettes.R
\name{hcl_palettes}
\alias{hcl_palettes}
\alias{print.hcl_palettes}
\alias{summary.hcl_palettes}
\alias{plot.hcl_palettes}
\alias{qualitative_hcl}
\alias{sequential_hcl}
\alias{diverge_hcl}
\title{HCL Color Palettes}
\usage{
hcl_palettes(type = NULL, name = NULL, plot = FALSE, n = 5L, ...)

\method{print}{hcl_palettes}(x, ...)

\method{summary}{hcl_palettes}(object, ...)

\method{plot}{hcl_palettes}(x, n = 5L, fixup = TRUE, ...)

qualitative_hcl(n, h = c(0, 360 * (n - 1)/n), c = 50, l = 70,
  fixup = TRUE, alpha = 1, palette = NULL, rev = FALSE, ..., h1, h2, c1,
  l1)

sequential_hcl(n, h = 260, c = 80, l = c(30, 90), power = 1.5,
  gamma = NULL, fixup = TRUE, alpha = 1, palette = NULL, rev = FALSE,
  ..., h1, h2, c1, c2, l1, l2, p1, p2, c.)

diverge_hcl(n, h = c(260, 0), c = 80, l = c(30, 90), power = 1.5,
  gamma = NULL, fixup = TRUE, alpha = 1, palette = NULL, rev = FALSE,
  ..., h1, h2, c1, l1, l2, p1, p2)
}
\arguments{
\item{type}{character indicating type of HCL palette.}

\item{name, palette}{character. Name of HCL color palette.}

\item{plot}{logical. Should the selected HCL color palettes be visualized?}

\item{n}{the number of colors (\eqn{\ge 1}{>= 1}) to be in the palette.}

\item{\dots}{Other arguments passed to \code{\link{hex}}.}

\item{x, object}{A \code{hcl_palettes} object.}

\item{fixup}{logical. Should the color be corrected to a valid RGB value
before correction?}

\item{h, h1, h2}{hue value in the HCL color description, has to be in [0, 360].}

\item{c, c., c1, c2}{chroma value in the HCL color description.}

\item{l, l1, l2}{luminance value in the HCL color description.}

\item{alpha}{numeric vector of values in the range \code{[0, 1]} for alpha
transparency channel (0 means transparent and 1 means opaque).}

\item{rev}{logical. Should the color palette vector be returned in reverse order?}

\item{power, p1, p2}{control parameter determining how chroma and luminance should
be increased (1 = linear, 2 = quadratic, etc.).}

\item{gamma}{Deprecated.}
}
\description{
Qualitative, sequential (single-hue and multi-hue), and diverged
color palettes based on the HCL (hue-chroma-luminance) color model.
}
\details{
Details are still tbd...
}
\examples{
## overview of all _named_ HCL palettes
hcl_palettes()

## visualize
hcl_palettes("qualitative", plot = TRUE)
hcl_palettes("sequential (single-hue)", plot = TRUE)
hcl_palettes("sequential (multi-hue)", plot = TRUE)
hcl_palettes("diverging", plot = TRUE)

## inspect a specific palette
## (upper-case, spaces, etc. are ignored for matching)
hcl_palettes(name = "Dark 2")
hcl_palettes(name = "dark2")

## set up actual colors
qualitative_hcl(4, h = c(0, 288), c = 50, l = 60) ## by hand
qualitative_hcl(4, palette = "dark2")             ## by name
qualitative_hcl(4, palette = "dark2", c = 80)     ## by name plus modification
}
\keyword{color}
