% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lighten.R
\name{lighten}
\alias{lighten}
\alias{darken}
\title{Lighten or Darken Colors by Modifying Luminance in HCL Space}
\usage{
lighten(col, amount = 0.1, method = "relative", space = c("HLS", "HCL"),
  fixup = TRUE)

darken(col, amount = 0.1, ...)
}
\arguments{
\item{col}{vector of any of the three kind of R colors, i.e., either a color
name (an element of \code{\link[grDevices]{colors}}), a hexadecimal string
of the form \code{"#rrggbb"} or \code{"#rrggbbaa"} (see
\code{\link[grDevices]{rgb}}), or an integer \code{i} meaning
\code{palette()[i]}.}

\item{amount}{numeric specifying the amount of lightening. This is applied either
multiplicatively or additively to the luminance value, depending on the
setting of \code{method} (either relative or absolute). Negative numbers
cause darkening.}

\item{...}{Other parameters handed to the function \code{lighten()}.}
}
\value{
A character vector with (s)RGB codings of the colors in the palette.
}
\description{
Transform a vector of given colors to the corresponding colors with luminance increased or 
decreased in HCL space. When darkening colors, chroma may be decreased as well if needed.
}
\details{
The given colors are first transformed to RGB (either using \code{\link[colorspace]{hex2RGB}} or
\code{\link[grDevices]{col2rgb}}) and then to HCL (\code{\link[colorspace]{polarLUV}}).
In HCL, luminance is adjusted and then the color is transformed back to a hexadecimal
string.

The function \code{lighten()} ligthens colors by either proportionally or absolutely increasing
the luminance L. The matehmatical formula used to transform L is \code{(1 + amount) * L} if
\code{method = "relative"}, and \code{L + amount} if \code{method = "absolute"}. If \code{amount}
is negative then colors are darkened rather than lightened. The function \code{darken()} is a convenience
function that multiplies \code{amount} with -1 and then calls \code{lighten()}.
}
\examples{
## rainbow of colors and their desaturated counterparts
rainbow_hcl(12)
lighten(rainbow_hcl(12))

## convenience demo function
wheel <- function(col, radius = 1, ...)
  pie(rep(1, length(col)), col = col, radius = radius, ...) 

## compare original, lightened, and darkened colors
par(mar = rep(0, 4), mfrow = c(2, 3))
## rainbow color wheel
wheel(rainbow_hcl(12))
wheel(lighten(rainbow_hcl(12), amount = 0.1))
wheel(lighten(rainbow_hcl(12), amount = 0.3)) 
wheel(rainbow_hcl(12))
wheel(darken(rainbow_hcl(12), amount = 0.1))
wheel(darken(rainbow_hcl(12), amount = 0.3))

}
\seealso{
\code{\link[colorspace]{polarLUV}}, \code{\link[colorspace]{hex}}, \code{\link[colorspace]{desaturate}}
}
\keyword{color}
