% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lighten.R
\name{lighten}
\alias{lighten}
\alias{darken}
\title{Algorithmically Lighten or Darken Colors}
\usage{
lighten(col, amount = 0.1, method = c("relative", "absolute"),
  space = c("HCL", "HLS", "combined"), fixup = TRUE)

darken(col, amount = 0.1, space = "combined", ...)
}
\arguments{
\item{col}{vector of any of the three kind of R colors, i.e., either a color
name (an element of \code{\link[grDevices]{colors}}), a hexadecimal string
of the form \code{"#rrggbb"} or \code{"#rrggbbaa"} (see
\code{\link[grDevices]{rgb}}), or an integer \code{i} meaning
\code{palette()[i]}.}

\item{amount}{numeric specifying the amount of lightening. This is applied either
multiplicatively or additively to the luminance value, depending on the
setting of \code{method} (either relative or absolute). Negative numbers
cause darkening.}

\item{method}{character string specifying the adjustment method. Can be either \code{"relative"} or \code{"absolute"}.}

\item{space}{character string specifying the color space in which adjustment happens. Can be either \code{"HLS"} or \code{"HCL"}.}

\item{fixup}{logical If set to \code{TRUE}, colors that fall outside of the RGB color gamut are slightly modified
by translating individual primary values so they lie between 0 and 255. If set to \code{FALSE}, out-of-gamut colors
are replaced by \code{NA}.}

\item{...}{Other parameters handed to the function \code{lighten()}.}
}
\value{
A character vector with (s)RGB codings of the colors in the palette.
}
\description{
The functions \code{lighten} and \code{darken} take a vector of R colors and adjust the colors such that
they appear lightened or darkened, respectively.
}
\details{
The color adjustment can be calculated in three different color spaces. First, if \code{space = "HCL"},
the colors are transformed to HCL, (\code{\link[colorspace]{polarLUV}}), the luminance component L
is adjusted, and then the colors are transformed back to a hexadecimal RGB string. Second, if 
\code{space = "HLS"}, the colors are transformed to HLS, the lightness component L is adjusted, and
then the color is transformed back to a hexadecimal RGB string. Finally, if \code{space = "combined"},
the colors are first adjusted in both the HCL and HLS spaces. Then, the adjusted HLS colors are converted
into HCL, and then the chroma components of the adjusted HLS colors are copied to the adjusted HCL colors.
Thus, in effect, the combined model adjusts luminance in HCL space but chroma in HLS space. We have found
that, in general, \code{space = "HCL"} performs best for lightening colors and \code{space = "combined"}
performs best for darkening colors, and these are the default settings for \code{lighten} and \code{darken},
respectively.

Regardless of the chosen color space, the adjustment of the L component can occur by two methods, relative
(the default) and absolute. Under the relative method, the adjustment is \code{100 - (100 - L) * (1 - amount)} when
lightening colors and \code{L * (1 - amount)} when darkening colors. Under the absolute method, the adjustment
is \code{L +/- 100 * amount} when lightening/darkening colors.

Programmatically lightening and darkening colors can yield unexpected results (see examples). In HCL space,
colors can become either too gray or overly colorful. By contrast, in HLS space it can happen that the
overallamount of lightening or darkening appears to be non-uniform among a group of colors that are 
lightened or darkened jointly, and again, colors can become either too gray or overly colorful. We
recommend to try different color spaces if the default space for the chosen function (\code{lighten} 
or \code{darken}) does not look right in a specific application.
}
\examples{
## convenience demo function
pal <- function(col, border = "light gray") {
  n <- length(col)
  plot(0, 0, type="n", xlim = c(0, 1), ylim = c(0, 1), axes = FALSE,
       xlab = "", ylab = "")
  rect(0:(n-1)/n, 0, 1:n/n, 1, col = col, border = border)
}

# lighten dark colors, example 1
cl <- rainbow_hcl(5)
par(mfrow = c(3, 3), mar = rep(0, 4), oma = c(0, 0, 2, 0))
pal(cl); mtext("HCL")
pal(cl); mtext("HLS")
pal(cl); mtext("combined")
pal(lighten(cl, 0.15))
pal(lighten(cl, 0.15, space = "HLS"))
pal(lighten(cl, 0.15, space = "combined"))
pal(lighten(cl, 0.3))
pal(lighten(cl, 0.3, space = "HLS"))
pal(lighten(cl, 0.3, space = "combined"))

# lighten dark colors, example 2
cl <- c("#61A9D9", "#ADD668", "#E6D152", "#CE6BAF", "#797CBA")
par(mfrow = c(3, 3), mar = rep(0, 4), oma = c(0, 0, 2, 0))
pal(cl); mtext("HCL")
pal(cl); mtext("HLS")
pal(cl); mtext("combined")
pal(lighten(cl, 0.15))
pal(lighten(cl, 0.15, space = "HLS"))
pal(lighten(cl, 0.15, space = "combined"))
pal(lighten(cl, 0.30))
pal(lighten(cl, 0.30, space = "HLS"))
pal(lighten(cl, 0.30, space = "combined"))

# darken light colors, example 1
cl <- rainbow_hcl(5, l = 90, c = 35)
par(mfrow = c(3, 3), mar = rep(0, 4), oma = c(0, 0, 2, 0))
pal(cl); mtext("combined")
pal(cl); mtext("HCL")
pal(cl); mtext("HLS")
pal(darken(cl, 0.15))
pal(darken(cl, 0.15, space = "HCL"))
pal(darken(cl, 0.15, space = "HLS"))
pal(darken(cl, 0.30))
pal(darken(cl, 0.30, space = "HCL"))
pal(darken(cl, 0.30, space = "HLS"))

# darken light colors, example 2 
cl <- c("#CDE4F3","#E7F3D3","#F7F0C7","#EFCFE5","#D0D1E7")
par(mfrow = c(3, 3), mar = rep(0, 4), oma = c(0, 0, 2, 0))
pal(cl); mtext("combined")
pal(cl); mtext("HCL")
pal(cl); mtext("HLS")
pal(darken(cl, 0.15))
pal(darken(cl, 0.15, space = "HCL"))
pal(darken(cl, 0.15, space = "HLS"))
pal(darken(cl, 0.30))
pal(darken(cl, 0.30, space = "HCL"))
pal(darken(cl, 0.30, space = "HLS"))
}
\seealso{
\code{\link[colorspace]{polarLUV}}, \code{\link[colorspace]{hex}}, \code{\link[colorspace]{desaturate}}
}
\keyword{color}
